+++ UVtools.Core/FileFormats/PHZFile.cs
@@ -943,26 +943,21 @@ namespace UVtools.Core.FileFormats
 
         public override float PrintTime
         {
-            get => HeaderSettings.PrintTime;
+            get => base.PrintTime;
             set
             {
-                HeaderSettings.PrintTime = (uint) value;
                 base.PrintTime = value;
+                HeaderSettings.PrintTime = (uint)base.PrintTime;
             }
         }
 
         public override float MaterialMilliliters
         {
-            get
-            {
-                var materialMl = base.MaterialMilliliters;
-                return materialMl > 0 ? materialMl : HeaderSettings.VolumeMl;
-            }
+            get => base.MaterialMilliliters;
             set
             {
-                if (value <= 0) value = base.MaterialMilliliters;
-                HeaderSettings.VolumeMl = (float)Math.Round(value, 3);
-                base.MaterialMilliliters = HeaderSettings.VolumeMl;
+                base.MaterialMilliliters = value;
+                HeaderSettings.VolumeMl = base.MaterialMilliliters;
             }
         }